requirements.txt
requests==2.28.2
smtplib
lxml==4.9.3
python-dotenv
